datastore_service/requirements.txt
pandas
numpy

psycopg2-binary
sqlalchemy>=1.4.0
sqlalchemy-utils
alembic

asyncpg
databases[postgresql]

aiohttp
contextlib2

pyyaml
python-dotenv

beautifulsoup4
openpyxl
lxml
html5lib

yahooquery
alpaca_trade_api

luigi
